refactor: BoardViewsController 객체 생성 시 초기화 로직 수정
- 객체 생성 시 db에 저장된 돌들을 오목판에 세팅
- 객체 생성 시 보드뷰에 클릭 리스너 세팅 (클릭 시 돌을 놓는 리스너)
