+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -7,11 +7,6 @@ import android.widget.TableLayout
 import android.widget.TableRow
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.children
-import domain.Board
-import domain.stone.BlackStone
-import domain.stone.Stone
-import domain.stone.Stones
-import domain.stone.WhiteStone
 import woowacourse.omok.database.OmokDB
 
 
@@ -31,32 +26,12 @@ class MainActivity : AppCompatActivity() {
             .filterIsInstance<ImageView>().toList()
 
         omokDB = OmokDB(this)
-        boardViewsController = BoardViewsController(this, boardViews)
-
-        initBoardView()
-        boardViewsController.setPutStoneOnClickListener(omokDB)
-
+        boardViewsController = BoardViewsController(this, boardViews, omokDB)
 
         val restartButton = findViewById<Button>(R.id.button_restart)
         restartButton.setOnClickListener {
             boardViewsController.resetView()
             omokDB.deleteDB()
-            boardViewsController.board = Board(Stones(setOf()))
         }
-    }
-
-    private fun initBoardView() {
-        //db에 저장된 돌들을 오목판에 세팅
-        val dbStones = omokDB.getStoneData()
-        val stones = mutableSetOf<Stone>()
-        dbStones.forEach {
-            val (omokPoint, stoneColor) = it
-            boardViewsController.setDBStoneOnView(omokPoint, stoneColor)
-            val (x, y) = omokPoint
-            val stone: Stone =
-                if (stoneColor == StoneColor.BLACK.english) BlackStone(x, y) else WhiteStone(x, y)
-            stones.add(stone)
-        }
-        boardViewsController.board = Board(Stones(stones))
     }
 }
